Revert "Dev (#16)"
This reverts commit 4245a67ece792143a1973c6e680ef5411cb1f258.

+++ src/main/java/com/hims/response/BillingPolicyResponse.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @AllArgsConstructor
@@ -16,7 +15,7 @@ public class BillingPolicyResponse {
     private String description;
     private String applicableBillingType;
     private Integer followupDaysAllowed;
-    private BigDecimal discountPercentage;
+    private Integer discountPercentage;
     private String createdBy;
     private String lastUpdatedBy;
     private LocalDateTime lastUpdateDate;
